python: avoid unlikely sign extension bug
Detected by Coverity; same analysis as for commit f73198df.

* python/libvirt-override.c (libvirt_virDomainGetVcpuPinInfo): Use
correct type.

+++ python/libvirt-override.c
@@ -1114,7 +1114,7 @@ libvirt_virDomainGetVcpuPinInfo(PyObject *self ATTRIBUTE_UNUSED,
     virNodeInfo nodeinfo;
     virDomainInfo dominfo;
     unsigned char *cpumaps;
-    int cpumaplen, vcpu, pcpu;
+    size_t cpumaplen, vcpu, pcpu;
     unsigned int flags;
     int i_retval;
